PY4E Following Links in HTML Using BeautifulSoup
Assignment: Following Links in HTML Using BeautifulSoup.
In this assignment you will write a Python program that expands on https://www.py4e.com/code3/urllinks.py. The program will use urllib to read the HTML from the data files below, extract the href= values from the anchor tags, scan for a tag that is in a particular position from the top and follow that link, repeat the process a number of times, and report the last name you find.